refactor: menuGroup 리팩토링
