+++ src/main/java/kitchenpos/menus/application/MenuService.java
@@ -2,7 +2,7 @@ package kitchenpos.menus.application;
 
 import kitchenpos.menus.application.dto.MenuCreateRequest;
 import kitchenpos.menus.application.dto.MenuPriceChangeRequest;
-import kitchenpos.menus.domain.*;
+import kitchenpos.menus.tobe.domain.MenuGroupRepository;
 import kitchenpos.menus.tobe.domain.MenuRepository;
 import kitchenpos.menus.tobe.domain.Menu;
 import kitchenpos.menus.tobe.domain.PurgomalumClient;
