+++ src/main/java/kitchenpos/menus/application/dto/MenuGroupCreateRequest.java
@@ -1,0 +1,13 @@
+package kitchenpos.menus.application.dto;
+
+public class MenuGroupCreateRequest {
+    private final String name;
+
+    public MenuGroupCreateRequest(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
