+++ src/main/java/kitchenpos/menus/infra/repository/JpaMenuGroupRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.menus.infra.repository;
 
-import kitchenpos.menus.domain.MenuGroup;
-import kitchenpos.menus.domain.MenuGroupRepository;
+import kitchenpos.menus.tobe.domain.MenuGroup;
+import kitchenpos.menus.tobe.domain.MenuGroupRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
