+++ src/main/java/kitchenpos/menus/tobe/domain/MenuGroup.java
@@ -1,9 +1,10 @@
-package kitchenpos.menus.domain;
+package kitchenpos.menus.tobe.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import java.util.Objects;
 import java.util.UUID;
 
 @Table(name = "menu_group")
@@ -16,22 +17,26 @@ public class MenuGroup {
     @Column(name = "name", nullable = false)
     private String name;
 
+    public static MenuGroup newOne(String name) {
+        if (Objects.isNull(name) || name.isEmpty()) {
+            throw new IllegalArgumentException();
+        }
+        return new MenuGroup(UUID.randomUUID(), name);
+    }
+
     public MenuGroup() {
+    }
+    public MenuGroup(UUID id, String name) {
+        this.id = id;
+        this.name = name;
     }
 
     public UUID getId() {
         return id;
     }
 
-    public void setId(final UUID id) {
-        this.id = id;
-    }
-
     public String getName() {
         return name;
     }
 
-    public void setName(final String name) {
-        this.name = name;
-    }
 }
