+++ src/main/java/kitchenpos/menus/tobe/domain/MenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.domain;
+package kitchenpos.menus.tobe.domain;
 
 import java.util.List;
 import java.util.Optional;
