+++ src/test/java/kitchenpos/MenuFixture.java
@@ -1,6 +1,6 @@
 package kitchenpos;
 
-import kitchenpos.menus.domain.MenuGroup;
+import kitchenpos.menus.tobe.domain.MenuGroup;
 import kitchenpos.menus.tobe.domain.*;
 
 import java.math.BigDecimal;
@@ -29,10 +29,7 @@ public class MenuFixture {
     }
 
     public static MenuGroup menuGroup(final String name) {
-        final MenuGroup menuGroup = new MenuGroup();
-        menuGroup.setId(UUID.randomUUID());
-        menuGroup.setName(name);
-        return menuGroup;
+        return new MenuGroup(UUID.randomUUID(), name);
     }
 
     public static MenuProduct menuProduct() {
